Rename TextTransformer into ReactTextTransformer
