+++ src/transform-by-pattern.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { Fragment, ReactElement } from 'react'
-import { SingleTextLike, TextLike, TextMatcherPattern } from './text-transformer'
+import { SingleTextLike, TextLike, TextMatcherPattern } from './react-text-transformer'
 import { mergeAdjacentStrings } from './utils'
 
 function wasElementProcessedAndBlocked(element: ReactElement<{ children: TextLike }>): boolean {
